refactor: Split Firebase module into separate files to force chunk regeneration
Completely restructure Firebase initialization by splitting into three files:
- config.ts: Environment variable validation and configuration
- initialize.ts: Singleton initialization logic and service getters
- index.ts: Lazy-loaded Proxy exports

This radical restructuring will force webpack to generate entirely new
chunk hashes, breaking the CDN cache collision that was serving old
chunks with undefined environment variables.
